requirements.txt
# Core
numpy>=1.24
pandas>=2.0
pyyaml>=6.0
tqdm>=4.66
pillow>=10.0
rich>=13.7

# ML + metrics
scikit-learn>=1.3

# PyTorch stack (install correct CUDA build separately if needed)
torch>=2.1
torchvision>=0.16

# CLIP
open-clip-torch>=2.24.0

# Hugging Face datasets
datasets>=2.18
huggingface-hub>=0.20

# Plotting
matplotlib>=3.7

# Optional: demo UI
# gradio>=4.0

# Optional: DIRE branch (enable only when you implement it)
# diffusers>=0.25
# transformers>=4.38
# accelerate>=0.26
# safetensors>=0.4

# Optional: dev/test tooling
# pytest>=7.4
# ruff>=0.2
